v1.0

+++ rss.pl
@@ -10,30 +10,40 @@
 
 use LWP::Simple;
 
-$infile = "./urls.list";		# The temp file that holds the rss feed
-$download = "./downloaded";		# The file that holds the links for the previously downloaded shows
-@rss_link = ('http://revision3.com/systm/feed/quicktime-high-definition/', 'http://revision3.com/tekzilla/feed/quicktime-high-definition/?subshow=false', 'http://revision3.com/hak5/feed/quicktime-large/'); # Array holding RSS links
+$debug = 0; # set to 1 to stop program actually downloading also does not delete urls.list file
 
+$infile = "./urls.list";	# The temp file that holds the rss feed
+$feedsfile = 'feeds'; 		# The file that contains all the feed links and names
+$destination = '/test/'		# The Destination to save fils
+open(FEEDS, $feedsfile);	# open feedsfile for reading feed info
+@lines = <FEEDS>;
 
-#$rss_link = 'http://revision3.com/systm/feed/quicktime-high-definition/'; # Array holding RSS links
+foreach $line(@lines)		# foreach loop to split the feeds file format = NAME,FEEDURL
+{
+ @temp = split(/,/, $line);
+ push(@names, $temp[0]);  	# store the name of the feed
+ push(@rss_link, $temp[1]);	# store the link for the rss feed
+}
 
+$regex = 'enclosure url="(http://.*\.mp4)"'; 	# RegEx searching for shows in .mp4 format
+my @matches; 	# Array to hold the details of matches from the RegEx
 
-$regex = 'enclosure url="(http://.*\.mp4)"'; # RegEx searching for shows in .mp4 format
-my @matches; # Array to hold the details of matches from the RegEx
+$counter = 0;
 
 foreach $rss_link(@rss_link)
 {
  print $rss_link."\n";
- getstore("$rss_link","$infile"); # download the rss feed for today.
-
+ getstore("$rss_link","$infile"); 	# download the rss feed for today.
+ 
+ $downloaded = './downloaded/'.$names[$counter];
  open(INFILE, $infile);		# Open the temp file
- open(DOWNLOAD,$download);	# Open the previously downloaded file (Read-Only)
+ open(DOWNLOAD,$downloaded);	# Open the previously downloaded file (Read-Only)
 
  @prevdown = <DOWNLOAD>;		# store the previously downloaded files 
 #print "Ignoring $ignorenum files previously downloaded($download): \n @prevdown"; 
 
  close(DOWNLOAD); 	# Cloes the read-only status on the download file
- open(DOWNLOAD,">>$download"); 	# Open with appened
+ open(DOWNLOAD,">>$downloaded"); 	# Open with appened
  
  while ($line = <INFILE>)
  {
@@ -52,7 +62,10 @@ foreach $rss_link(@rss_link)
     {
      print DOWNLOAD $1."\n";		# Print link to downloaded file
      print "Download:".$1."\n";			# Print linke to screen
-     exec ('pushd /test/ && wget '.$1."&& popd"); # command = move to /test/, execute "wget link", move back to previous folder
+     if ($debug == 0)
+     {
+      exec ("pushd $destination && wget ".$1."&& popd"); # command = move to /test/, execute "wget link", move back to previous folder
+     }
     } #if $temp
    
   } # if $linke $regex
@@ -60,5 +73,9 @@ foreach $rss_link(@rss_link)
 
  close(INFILE);		# Close the file
  close(DOWNLOAD);
- unlink($infile); 	# Delete urls.list
+ if ($debug == 0)
+ {
+  unlink($infile); 	# Delete urls.list
+ }
+ $counter++;
 }
